trading-bot/requirements.txt
tradovate-api
pandas
numpy
requests
python-dotenv
matplotlib
scipy